+++ sysdeps/ieee754/flt-32/e_coshf.c
@@ -77,7 +77,7 @@ __ieee754_coshf (float x)
   double rm = sm * (te - h * to);
   double r = rp + rm;
   float ub = r;
-  double lb = r - 1.45e-10 * r;
+  float lb = r - 1.45e-10 * r;
   if (__glibc_unlikely (ub != lb))
     {
       const double iln2h = 0x1.7154765p+5;
